Added Inventory Controller
Inventory controller contains a Dictionary that has a key which is an Enum and a value of int.
The Enum will contain the name of all items that exist in the game and other scripts will use the dictionary and Enum to get the amount of each item the player has.