Added RiverHouse prefab, technically MVP works
Added a river house prefab at the end. If you wanna put stuff like narrative here, go for it, don't let me slow you down!!!

In a nutshell, level has a beginning and an end, no middle yet (the actual gameplay bit). But, in a worst case scenario, those basics are in