Inventory controller complete
Scripts can now ask how many items the inventory has, if the requested item doesn't exist then 0 is returned.
When adding items it checks if the item is in the dictionary, if not then it adds it.
When removing items it ensures that there are enough items before proceeding.

Both adding and removing items signal events that other scripts can subscribe themselves too.