Prototype dialogue implementation (#26)
* Merge main into branch (#24)

* Player traversal mechanics feedback (#21)

* Started implementing feedback

* Fixed Grapple Issues

Added Grapple Angle
Added Min Grapple Distance Push
Added Player Suicide Prevention
Limit Partition Camera Size

* Dash retains height

---------

Co-authored-by: MAX ANSELL <1903439@uad.ac.uk>

* River physics via Buoyancy Effectors

Added Water Body and Gravity Log prefabs. If the player stands on a Log, it acts as a platform but sinks a bit into the water. The player floats on the water. Currently not marked as a platform

---------

Co-authored-by: MAX ANSELL <1903439@uad.ac.uk>
Co-authored-by: HPetch <60263607+HPetch@users.noreply.github.com>

* Revert "Merge main into branch (#24)" (#25)

This reverts commit 7842c8728d06d0e6e6cef104488a94ca33a4dbf5.

---------

Co-authored-by: MAX ANSELL <1903439@uad.ac.uk>
Co-authored-by: HPetch <60263607+HPetch@users.noreply.github.com>