Decoration hover UI
UI for returning the decoration to the inventory and changing styles have now been added. UI disappears if the player isn't holding their mouse over the object for 1 second, or when picking the object up.

Functionally has been setup into empty functions waiting for the inventory system and decoration styles.