Trash complete
Trash now slows down as it reaches the breaking distance to look lite it is resisting the player pulling it. After reaching the correct distance a 0.5 second timer starts, if this timer completes then the trash is pulled.
If the player moves their mouse back within the break distance before the pull is complete then it is canceled.

The trash now checks if anything is infront of it before respawning, if the player or furniture is in the way then it doesn't respawn.