Merge main into branch (#44)
* Selector Rewrite

Mostly done rewriting the selector for the new system, just need to find the object that's under the mouse.

* Carrying on setting up furniture system

Player begins moving a DecorationObject
DecorationObject becomes greyed out
A new DecorationMovingObject is instantitated which copies the appearance of the DecorationObject
The DecorationMovingObject then follows the selector
If the placedown is cancelled then the DecorationMovingObject is lerped back to the DecorationObject then deleted, the DecorationObject is no loger greyed out
If placedown is done and it is valid the the DecorationMovingObject is deleted and the decoration object is moved to it's position

* Working on adding decoration object movement

Push before lunch break

* Unity Crash push

* Furniture system movement prototype complete

Furniture objects can now be moved, can only be placed if they are not overlapping with furniture or platforms. Attachment points must also be colliding with platforms.

Has many bugs and needs a lot of refactoring.

* Decoration system rotation and bug fixes

Fixed all known bugs for the placement of decoration objects, also now supports rotation of 5 degree increments between -20 and 20. Need to do some more refactoring before I consider it complete.

* Decoration system update

Starting on making the furniture system stand-alone so it can be added to main.

* Push post crash

Setting up UI to remove decorations and edit styles of decorations

* Decoration hover UI

UI for returning the decoration to the inventory and changing styles have now been added. UI disappears if the player isn't holding their mouse over the object for 1 second, or when picking the object up.

Functionally has been setup into empty functions waiting for the inventory system and decoration styles.

* Decoration object inheritance

Starting on reformatting decoration objects to revolve around inheritance. Will have base DecorationObject class which holds all general variables and functions, each type of object will then inherit this object type and add their bespoke features.

* Trash pull

Finished off inheritance system for decoration objects. Trash objects can be pulled to be destroyed. Vibrates when close to breaking. Need to make it not hacked together.

* Update InputManager.asset

added drop input

* Preparing to pull main

Backing up the progress I've made today to get ready to pull main into this branch. The inventory system is in main now so I must do this to progress with the decoration system.

* Ready to pull main

* Fixed wall collision issue

Wall collision was disabled by the furniture merge

* Editor Build Settings working (#41)

Build is setup for WebGL correctly for Itchio.
Important settings to know about in the Player Settings:

**Version** - typically just used in Analytics, but for cleaning sake and making sure Itch recognises updated files, set the number to the sprint we are on (e.g. 0.3.x = Sprint 3 build number x)

**Publishing Settings compression type** - Brotli does not work with anything, this should be set to Gzip. If issues with infinite loading screens still exist, try setting it to No Compression, but this is a last resort and not acceptable for gold master builds.

**Viewport size:** Make this a 16:9 resolution.

**WebGL Template:** Unknown at the moment, current set to Default, preferably we want to use Minimal or a Better Minimal template (not this sprint target)

* Player traversal improvements (#43)

* Merge main into recluse branch

* Test House given sample collision via sprite shape

Makes the character controller drift, but it works!

* Player sprite updated

Apparently this will be replaced/polished for week 6, but oh well

* Platform effectors fixed

* Player game feel updates

The player controller's values have been tweaked in many ways.

Jump height is now larger.
Movement is slightly slower but graceful.
Wall jumping can happen on wider away walls.
Each part of the mechanic no longer overshadows another.
Glide speed has decreased but the threshold of speed to trigger a glide has been doubled to avoid annoying accidental occurrences.

---------

Co-authored-by: LEWIS MACKAY <2002497@uad.ac.uk>
Co-authored-by: TIARNAN LAVERY <2002262@uad.ac.uk>
Co-authored-by: Kritso900 <69122361+Kritso900@users.noreply.github.com>
Co-authored-by: HPetch <60263607+HPetch@users.noreply.github.com>