Setup prefabs. Breaking resistance improvements. Sprite rotation while breaking. Elasticity improvements. Add items to inventory improvements.
Setup prefabs for both pickup object and pickup block object for easy creation of new objects.
Added a prefab for a 2 health bush and a 3 health bramble as examples of what a completed pickup object should look like.

Breaking resistance is now based on current health, so the higher health means the harder it is to break the object. Object now begins breaking at 0.95 max distance to prevent the funky math timings from the lerp. Extended the breaking timer so it is now 0.5 * health seconds, this works better with the new resistance and breaking system.

Objects now vibrate rotation once they've passed the breaking threshold to make an object breaking visually obvious.

If an object's target position is closer to it's origin than it's current position then ignores the slowdown from distance. Helps to give the object an elastic look.

Objects can now add items to inventory on each stage of damage, allows for things like pulling the leaves off a bush gives leaves then pulling the stem gives wood.