Made base sprite able to stay visible when inactive. Refactored and commented.
Base sprite can now be visible even when broken or inactive. If a 1 health pickup has a base then it will always be visible. If a higher health pickup object has a sprite in the final base sprite in the damage displayed sprites dictionary then it will be visible.

Refactored pickup object and added comments to all important lines.