First playable fix regressions (#67)
* Editor settings updated version number

* Removed UnityEditorInternal needless reference

* TextEffect.cs - remove UnityEditor ref and make vec2 declarations compile

new(1,1) was not syntax my version of VS was liking and I needed to add the Vector2 name in order for it to compile.

No, this doesn't fix the elusive issue sadly

* Added basic menu

I thought maybe the Awake functions weren't triggering correctly in the first scene of the build, so added a simple title screen.

Does not fix the build.

* Enabled Dev Build options for WebGL

Full stacktrace from WebGL will give us more info on what errors are happening, as well as the devlog. Just won't tell us exactly what lines are the issues.

I've gotten the issue reproducing on Windows when Dev Build is enabled. Reporting it on Jira

* Fixed TextEffect build bug

CharacterCanvas Awake was calling TextEffect ClearText before TextEffect had a chance to Initialise.

Fixed text box resizing

Made recluse alt colour have a non 0 alpha so it actually shows up

* Project settings updated, jiggle text removed for now

Jiggle text updates based on framerate or screen size or something, as it was unreadable in the demo text I was using

---------

Co-authored-by: Wolf1938 <38575320+Wolf1938@users.noreply.github.com>