Starting making craft menu
Furniture objects now know what crafting resources they contain.

Decoration Controller contains a dictionary of all furniture objects with a boolean of whether it can be accessed or not. Decoration Craft Menu then gets all furniture objects which are accessible and adds them to a list, this list is the objects that can be placed.

Currently can press 1 to create a furniture item, it checks if you have enough resources then if you have enough will remove them and spawn the object to the selector.