Added functionality to Harry's furniture craft UI
DecorationCraftMenu. Has 3 category tabs and 12 buttons. When clicking on a category tab it looks at all unlocked furniture objects and assigns the objects to the Decoration Craft Menu Buttons. When clicking on a button it shows it's name and crafting requirements on the right. Clicking "place" will close the menu and tell the Decoration Controller to spawn the object.

DecorationCraftMenuButton is attached to the buttons that contain the furniture objects, when changing tabs they hold the furniture objects and pass the held object to the Craft Menu when clicking them.

Decoration Objects now contain UiName, UiImage, UiFurnitureCatagory. These are used to display information in the DecorationCraftMenu.