Open and closing inventory via LeanTween
Different ease algorithms can be selected by dropdowns in the editor. You can't sequence break the opening and closing as they move to set values, not relative positions. As I'm using a recttransform, I can confirm it works across different resolutions (not in world units).

In DecorationCraftMenu you can use the OpenInventory() and CloseInventory() functions when you have selected an object and need to close the menu. It will handle it for you.