Fix: Images in crafting menu enforce aspect ratio
"Preserve Aspect" is an option that can be enabled for Images that will... preserve the sprite's aspect ratio.

This option is hidden in Debug unless a sprite is given to the Image. I turned on those options so we wouldn't have to do it through code for no reason.
Thanks Unity.