Tree level controller basic shell
I haven't set up TreeLevelInterface.cs as an Instance, sorry.

You call the Open and Close functions in TreeLevelInterface to trigger the open/close animation.

You can update text fields with UpdateMaterialText, by passing in the material type you want to update, and the integer value to update it to.