Tree level UI basic (#76)
* Fix: Images in crafting menu enforce aspect ratio

"Preserve Aspect" is an option that can be enabled for Images that will... preserve the sprite's aspect ratio.

This option is hidden in Debug unless a sprite is given to the Image. I turned on those options so we wouldn't have to do it through code for no reason.
Thanks Unity.

* Tree level controller basic shell

I haven't set up TreeLevelInterface.cs as an Instance, sorry.

You call the Open and Close functions in TreeLevelInterface to trigger the open/close animation.

You can update text fields with UpdateMaterialText, by passing in the material type you want to update, and the integer value to update it to.