0.4.3 Build - Debris float to corner, material count removed on furniture create, furniture first time place movement changed, idle sprite
If debris is a crafting material then moves to the top right, quest items still travel to player.
Numbers in top right change when the items reach the ui, but internally the inventory is modified instantly.
Material from the ui is removed when creating furniture, but planning to have materials move from the top right into the furniture to create it.

When creating furniture the furniture object now follows the move fake, after first placement only the move fake is moved.

Added idle sprite into game, currently plays when jumping at a stand still but will be fixed when we have a jumping sprite.

Can no longer create new furniture while holding furniture.

When the player doesn't have enough resources to craft the selected furniture the place button is now grayed out.