BGM and BGA can be controlled
Call the methods PlayMusic() or PlayAmbience() to play them. If you want to mute an audio source for whatever reason I will need to add more functionality to access that audio source's Mute boolean.

It affects the volume of the audio sources itself for fading, but *not* the mixer group. Mess with the mixer groups' volumes for controls via settings menu!