Audio bug fixed, removed old decoration prefabs, added tree level text, decoration selector improvements, fixed furniture cancel bug, made all furniture placement more forgiving, sped up debris, fixed pickup item randomisation
Functions in audio controller now don't error out when called without a sound.

Furniture and pickup prefabs that are no longer used have been removed.

The tree level UI now has text that clearly says the current tree level, this is likely to be removed in the future.

The decoration selector is no longer created and destroyed with edit mode, it is now permanent, this allows pickup objects to be pulled outside of edit mode. This required fixing a bug that allowed the decoration selector the pickup furniture while not in edit mode.

Decoration selector now moves with a more elastic look, making it visually clear that the selector does not need to be dragged far to break the object. This has been coded for a while but I forgot to implement it and had to fix a bug with the target location moving too far.

Fixed a bug where crafting furniture than cancelling it before placing it would remove the exp resulting in a negative exp total.

Made all furniture placement more forgiving, reduced the size of furniture hitboxes and lowered attach points. Added a boolean that allows the furniture to check for only a single attach point for the furniture to be valid, this is useful for hanging plants and similar furniture.

Sped up debris movement speed, movement speed now increases exponentially as the debris travel.

Fixed pickup item randomisation, randomisation for integers requires the max number to be increased by 1 for randomisation to work. Also increased the minimum values for wood to allow for easier progression.