Added Platform Non-Walk Layer, Trunk and Warsan area now have platform walk and platform non-walk.
Platform Non-Walk layer interacts with the decoration system but does allow the player to move on it, or regain their jump.
Trunk walk objects have a platform collision which give the collisions the player will walk on, but has no sprite
Trunk non-walk objects have a platform non-walk collision, this has the sprite for the trunk and all collision that the player will not walk on such as walls and ceilings.