Improved level transitions, Area name signs, Made platforming easier, Fixed library box placement issue, Fixed resource UI sprites being flipped. (#120)
Improved transition between spawn room and trunk tutorial, trunk and Warsan entrance, Warsan entrance and Warsan house, Warsan House and Warsan platforming section. This was done by overlapping the level partitions making transitions look smoother, and allowing for the player to see into the next partition before switching.

Added signs that give areas in the level names and directions to them, names are subject to change. Warsan's house is Brushwood cottage, Warsan platforming area is called Shrubfold Ridge, and Lucus's area is called Timber ruins. Lucus's area will receive more names when his area is redesigned.

Made tutorial dash jump easier by shortening the gap slightly, Made jumping up to solid branch platforms easier to climb by adding a one-way platform at the end of them to allow for easier access to the branch.

Forgot to enable boolean on library box that made it so it only required one placement point to be valid, this and been enabled now.

When changing the tree level UI from the right side of the screen to the left the position of the sprites in the UI was not reversed and so they were backwards to their intended orientation.