Fixed Pickup Object bug and Decoration Selector displaying incorrectly, Fixed Furniture Object preventing leaving edit mode, adjusted Furniture attach points
Fixed bug where pickup objects couldn't be cancelled while holding them due to a scripting oversight which caused cancel pickup to be called every frame, this caused the Decoration selector to not recognise that it was holding a pickup object which preventing the Cancel Pull function from being called.

Furniture objects did not have a material to switch to once out of decoration mode which prevented decoration mode from ending, this caused an error which could softlock the player.

Slightly adjusted attach points on Furniture Objects to allow for easier placement.