Fixed double pickup bug, finished pickup logic
While fixing PC-450 I found that I forgot to finish the logic for picking up and putting down furniture objects so I completing this logic which fixed the issue.
Now checks if your holding an object before allowing the player to pickup another object, also cancels placing an object if the player lets go of left mouse in an invalid spot.