Added Furniture Placement Freeform collider layer, Warsan 1st quest now accepts all 3 hanging baskets, Hanging baskets now correctly line up when placed, Can no longer interact with decoration objects behind the crafting UI
Added Furniture Placement Freeform collider layer, allows furniture to be placed on it but has no player collisions. Used on Warsan's house as the platform collider is a thin one-way collider rather than filling the roof.

Code that checks if a hanging planter has been placed now has or statements that check for all 3 types of hanging planters.

Removed unnecessary attach points from the hanging planters to make them fit better when placed.

Decoration craft UI now checks if the mouse is over it when open, if it is the it stops the player from interacting with objects behind it. Can still place object behind it but would be too obtrusive if it stopped this.